ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/setting/SettingRepository.java
package com.shopme.admin.setting;

import java.util.List;
import java.util.Optional;
import javax.transaction.Transactional;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.jpa.repository.Modifying;
import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.CrudRepository;
import org.springframework.data.repository.PagingAndSortingRepository;
import com.shopme.common.entity.Product;
import com.shopme.common.entity.Setting;
import com.shopme.common.entity.SettingCategory;

public interface SettingRepository extends CrudRepository<Setting, Integer>{
	public List<Setting> findByCategory(SettingCategory category);

}
